PCHR-4473: PHP 7.2 compatibility updates
Starting from PHP 7.1, microseconds are now added to DateTime objects. This means that
some DateTime comparisons might start failing in case it was assumed that microseconds will
not be present. Here's one example what a comparison starting to fail after the upgrade:
https://3v4l.org/bjBhH

In StatsCache.php, the issue was fixed by comparing the timestamp representation of the
DateTime objects. The timestamps don't include microseconds, so the comparison works as before.

On LeavePeriodEntitlementTest, the assertEquals now compares the formatted date instead of
the DateTime objects.

On the HRJobContractTestTrait, the contacts array would only be initialized at the moment the
first item is added to it. Inside the createContacts() method, the first thing we do is to
count the number of contacts in the list by using the count() function. That function only
works with countable variables and, the first time it gets called the contacts array will be
null (meaning it is not countable) and the code will throw an error.

Phpunit4 doesn't work with PHP 7.2, so now we need to run the tests with phpunit5. There,
the getMock() method is not available anymore and the available alternatives are createMock() or
getMockBuilder(). In the specific case in this commit, we still needed to use the class constructor,
so the only option that could be used was getMockBuilder(), as createMock() prevents the constructor
from being called.