PCHR-4473: Menu style fixes
CiviCRM 5.12 merged the uk.squiffle.kam extension to core.
This introduced drastical changes to the main navigation menu,
including a whole different markup to the entire menu.

Shoreditch had to be updated to be compatible with this new menu.
This was done in https://github.com/civicrm/org.civicrm.shoreditch/pull/386.
The decision there, however, was to cutomize the menu as little as
possible. The PR linked above has more details about these decisions
and screenshots of how the menu look like after the fixes.

As a result, even after updating CiviHR to the Shoreditch version
compatible with the new menu, things would still look different from
what they looked like before the upgrade. To fix this, we're now
using the bootstrapcivihr extension to re-add some of the customizations
that have been removed from Shoreditch and also adding some specific
to CiviHR as so as to make the menu look as close as possible to
what it looked like before.

On the CSS level, most of the changes were related to updating
selectors to properly match the new ones on the new menu.

On the JS level, the `org.civicrm.bootstrapcivihr/js/main-menu.js`
file has been added to provide the dynamic behavior of the quicksearch
box. It is basically a copy of the file (with some adjustments)
with the same name that used to exist on Shoreditch and that was
removed on the PR previously mentioned.