PCHR-4473: Fix for HRJobContract attached file issues
In CiviCRM 5.11, to improve access to attached file, the URL for file downloads
has been updated. It now contains a hash with some site specific info, which
can be used to prevent people from downloading all the files from a site
simply by manually changing the file id.

This same URL is used in Job Contracts to allow people to download the files
attached to a contract. However, we had some custom code in place that was overriding
the URL generated by Civi and removing the generated hash, resulting in users not being
able to download files.

The reason the custom code was put in place is unknown as the commit where it was
introduced (c3828ff60bad7000557e63cbebfac5986873b5f4) doesn't have any useful information.
Looking at the code, we can see that the only reason we were dealing with the file URL
was to pass the "frontend" param to the url() method in case the uploaded file was an
image. As stated in the documentation for the URL method, the frontend param is only
applicable to Wordpress and Joomla. Since CiviHR runs only on Drupal, the whole custom
code was useless, so it was removed.

Finally, while inspecting the code, I realized that the route civicrm/hrjobcontract/file/display
was not being used anywhere, so I deleted it