PCHR-4473: Add the Contact.getStaff API
There are many places in Leave & Absences where it's necessary
to access a list of contacts/staff in the company. Until today,
this was done by using the Contact.get API, but this has a few
problems:

- Due to ACLs, the API would only return contacts the current logged in
contact has access to. So, users that only have accees to their own
contact info would not be able to see other people on the leave calendar
- The Contact API returns a lot of info that is not really necessary
for leave and absences (like emails, phone numbers and addresses)

To avoid these two issues, the Contact.getStaff API is introduced.
Basically, it delegates to Contact.get API, but it does two important
things while doing that:

- It forces check_permissions to always be false, which results in
ACLs not being applied.
- It makes sure that only very basic info about the contacts will
be included in the response. That is, it will only return a limited
set of hardcoded fields: id, first, middle, last and display names.

Note that this limited set is applied even on the filters, meaning
it's not possible to the results by one field that is not included
in the response.

This commit also introduces some initial changes to the L&A angular
app so that it uses the new API instead of Contact.get, but this is
not complete and all the necessary changes will be done in a separate
commit.